+++ changelog.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.21] - 2020-02-13
+
+- Added 4 new items to page [/sea-tiger-newspapers](https://codyburleson.com/sea-tiger-newspapers)
+- Modified page [/sea-tiger-newspapers](https://codyburleson.com/sea-tiger-newspapers); added new found PDF, Vol III, No 95
+
 ## [1.5.20] - 2020-02-12
 
 - Modified page [/quotes-for-writers](https://codyburleson.com/quotes-for-writers); added 1 new quote by Will Self.